Add menu functionality - Extract DodajPostacJFrame from PostaciCreatorJPanel - update usages of DodajPostacJFrame - Add menu functionality in View - Add TextFilter used for choosing file
