+++ project/src/PostaciCreator/View/PostaciCreatorJPanel.java
@@ -1,22 +1,11 @@
 package PostaciCreator.View;
 
-import Bron.Lowcy.Luk;
-import Bron.Maga.Rozdzka;
-import Bron.Woja.BronWoja;
-import Bron.Woja.Miecz;
-import Bron.Woja.Mlot;
-import Bron.Zabojcy.Sztylet;
-import GraPackage.WyborKlasy;
 import PostaciCreator.Controller.Controller;
 import PostaciCreator.Controller.ControllerInterface;
 import PostaciCreator.Model.ModelInterface;
 import PostaciCreator.Model.Observers.ModelObservable;
 import PostaciCreator.Model.Observers.ModelObserver;
-import Postacie.Dystansowe.Lowca;
-import Postacie.Dystansowe.Mag;
 import Postacie.Postac;
-import Postacie.WZwarciu.Wojownik;
-import Postacie.WZwarciu.Zabojca;
 
 import javax.swing.*;
 import javax.swing.text.JTextComponent;
@@ -70,7 +59,7 @@ public class PostaciCreatorJPanel extends JPanel implements ModelObserver {
 		JButton zapiszButton = new JButton("Zapisz");
 		JButton wyjdzButton = new JButton("Wyjdz");
 
-		dodajButton.addActionListener(e -> new DodajPostacJFrame());
+		dodajButton.addActionListener(e -> new DodajPostacJFrame(controller));
 		zapiszButton.addActionListener(new Controller.ZapiszZmianyListener(model));
 		wyjdzButton.addActionListener(e -> parent.setVisible(false));
 
@@ -139,7 +128,7 @@ public class PostaciCreatorJPanel extends JPanel implements ModelObserver {
 			JButton editButton = new JButton("\\\\\\");
 			JButton deleteButton = new JButton("///");
 
-			editButton.addActionListener(e -> new DodajPostacJFrame(index, postac));
+			editButton.addActionListener(e -> new DodajPostacJFrame(index, postac, controller));
 			deleteButton.addActionListener(e -> controller.usunPostac(postac));
 
 			editButtons.add(editButton);
@@ -149,110 +138,8 @@ public class PostaciCreatorJPanel extends JPanel implements ModelObserver {
 		}
 	}
 
-	private class DodajPostacJFrame extends JFrame {
-		private final DefaultComboBoxModel<String> modelWoj = new DefaultComboBoxModel<>( WyborKlasy.listaBroniWojownika );
-		private final DefaultComboBoxModel<String> modelZab = new DefaultComboBoxModel<>( WyborKlasy.listaBroniZabojcy );
-		private final DefaultComboBoxModel<String> modelLow = new DefaultComboBoxModel<>( WyborKlasy.listaBroniLowcy );
-		private final DefaultComboBoxModel<String> modelMag = new DefaultComboBoxModel<>( WyborKlasy.listaBroniMaga );
-
-		private final int H_GAP = 5;
-		private final int V_GAP = 10;
-
-		private JPanel panel;
-		private JTextField imieField;
-		private JComboBox<String> klasaCB;
-		private JComboBox<String> bronCB;
-		private JButton zatwierdzButton;
-
-		private boolean editting;
-		private int index;
-
-		public DodajPostacJFrame(){
-			panel = new JPanel();
-			this.getContentPane().add(panel);
-			this.setSize(new Dimension(200, 200));
-			this.setDefaultCloseOperation(DO_NOTHING_ON_CLOSE);
-
-			imieField = new JTextField("Imie");
-
-			bronCB = new JComboBox<>(WyborKlasy.listaBroniWojownika);
-			bronCB.setSelectedIndex(0);
-
-			klasaCB = new JComboBox<>(WyborKlasy.listaKlas);
-			klasaCB.addActionListener(e -> {
-				switch (klasaCB.getSelectedIndex()){
-					case 0 -> bronCB.setModel(modelWoj);
-					case 1 -> bronCB.setModel(modelZab);
-					case 2 -> bronCB.setModel(modelLow);
-					case 3 -> bronCB.setModel(modelMag);
-				}
-			});
-			klasaCB.setSelectedIndex(0);
 
 
-			zatwierdzButton = new JButton("OK");
-			zatwierdzButton.addActionListener(e -> {
-				if (editting) controller.aktualizujPostac(stworzPostac(), index);
-				else controller.nowaPostac(stworzPostac());
-
-				this.setVisible(false);
-			});
-
-			panel.setLayout(new BoxLayout(panel, BoxLayout.Y_AXIS));
-			panel.add(imieField);
-			panel.add(Box.createRigidArea(new Dimension(H_GAP, V_GAP)));
-			panel.add(klasaCB);
-			panel.add(Box.createRigidArea(new Dimension(H_GAP, V_GAP)));
-			panel.add(bronCB);
-			panel.add(zatwierdzButton);
-			this.setVisible(true);
-		}
-
-		public DodajPostacJFrame(int index, Postac p){
-			this();
-			this.index = index;
-			editting = true;
-			imieField.setText(p.getImie());
-
-			int klasaIndex = 0;
-			switch (p.getClass().getSimpleName()){
-				case "Wojownik" -> klasaIndex = 0;
-				case "Zabojca" -> klasaIndex = 1;
-				case "Lowca" -> klasaIndex = 2;
-				case "Mag" -> klasaIndex = 3;
-			}
-			klasaCB.setSelectedIndex(klasaIndex);
-
-			int bronIndex = 0;
-			if (klasaIndex == 0 && p.getBron().getClass().getSimpleName().equals("Mlot")){
-				bronIndex = 1;
-			}
-			bronCB.setSelectedIndex(bronIndex);
-		}
-
-		private Postac stworzPostac(){
-			Postac p = null;
-			String imie = imieField.getText();
-
-			int index = (int)(Math.random() * WyborKlasy.listaImionBota.length);
-			if (imie.equals("")) imie += WyborKlasy.listaImionBota[index];
-
-			switch (klasaCB.getSelectedIndex()) {
-				case 0 -> {
-					BronWoja bronW;
-					if (bronCB.getSelectedIndex() == 0)
-						bronW = new Miecz();
-					else bronW = new Mlot();
-					p = new Wojownik(imie, bronW);
-				}
-				case 1 -> p = new Zabojca(imie, new Sztylet());
-				case 2 -> p = new Lowca(imie, new Luk());
-				case 3 -> p = new Mag(imie, new Rozdzka());
-			}
-			return p;
-		}
-
-	}
 
 	private void setupJTextComponentForView(JTextComponent comp){
 		comp.setEditable(false);
